backend/pyproject.toml
[project]
name = "django-vue-class"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "LucasGabrielAntonete", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=5.0.4",
    "setuptools>=69.5.1",
    "djangorestframework>=3.15.1",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.3.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
